Added files
**This is set up to be run as a simulation**

This is a project I have been working on as a part of a college class. This will take in the classes that a student has taken and put them into an easy-to-read GUI to show which classes have yet to be taken. I have already taken the liberty of setting up many of these files, so the only necessary file to run is test.py. At this current stage (Jan 21 2020) I have a few bugs I'm working on troubleshooting:

1) when a user successfully logs in, the original login window doesn't close automatically, and must manually be closed before the next GUI opens up. 
2) Upon re-declaring the listbox_widget the compiler throws an Attribute Error, stating that the 'int object has no attribute 'tk''. This doesn't stop the GUI from showing which classes are necessary for our user. 

Please note that these 'users' are based on characters from the TV show Archer, and the passwords are references to the show as well (can be found in the classes.py file)

This is still a work in progress at the moment, and I'll be working on this for the next few months or so to come.